requirements.txt
-e .
aws-cdk.aws-lambda==1.129.0
aws-cdk.aws-sam==1.129.0
aws-lambda-powertools